Move files from data/model to domain/model
