--- src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/data/model/AuthenticationMode.kt
@@ -1,19 +1,0 @@
-/* ******************************************************************************
- * Copyright (c) 2025 Calypso Networks Association https://calypsonet.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information
- * regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the BSD 3-Clause License which is available at
- * https://opensource.org/licenses/BSD-3-Clause.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- ****************************************************************************** */
-package org.calypsonet.keyple.demo.control.data.model
-
-enum class AuthenticationMode {
-  SAM,
-  PKI,
-  NO_AUTHENTICATION
-}
