--- src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/data/model/CardReaderResponse.kt
@@ -1,27 +1,0 @@
-/* ******************************************************************************
- * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information
- * regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the BSD 3-Clause License which is available at
- * https://opensource.org/licenses/BSD-3-Clause.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- ****************************************************************************** */
-package org.calypsonet.keyple.demo.control.data.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-import org.calypsonet.keyple.demo.control.domain.model.Status
-
-@Parcelize
-data class CardReaderResponse(
-    val status: Status,
-    val authenticationMode: AuthenticationMode,
-    val lastValidationsList: ArrayList<Validation>? = null,
-    val titlesList: ArrayList<Contract>,
-    val errorTitle: String? = null,
-    val errorMessage: String? = null
-) : Parcelable
