--- src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/data/model/Location.kt
@@ -1,24 +1,0 @@
-/* ******************************************************************************
- * Copyright (c) 2021 Calypso Networks Association https://calypsonet.org/
- *
- * See the NOTICE file(s) distributed with this work for additional information
- * regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the BSD 3-Clause License which is available at
- * https://opensource.org/licenses/BSD-3-Clause.
- *
- * SPDX-License-Identifier: BSD-3-Clause
- ****************************************************************************** */
-package org.calypsonet.keyple.demo.control.data.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Location(val id: Int, val name: String) : Parcelable {
-
-  override fun toString(): String {
-    return "$id - $name"
-  }
-}
