+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/domain/model/AppSettings.kt
@@ -1,7 +1,5 @@
 package org.calypsonet.keyple.demo.control.domain.model
 
-import org.calypsonet.keyple.demo.control.data.model.Location
-
 object AppSettings {
   lateinit var readerType: ReaderType
   lateinit var location: Location
