+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/domain/model/AuthenticationMode.kt
@@ -1,0 +1,7 @@
+package org.calypsonet.keyple.demo.control.domain.model
+
+enum class AuthenticationMode {
+  SAM,
+  PKI,
+  NO_AUTHENTICATION
+}
