+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/domain/model/CardReaderResponse.kt
@@ -1,0 +1,10 @@
+package org.calypsonet.keyple.demo.control.domain.model
+
+data class CardReaderResponse(
+    val status: Status,
+    val authenticationMode: AuthenticationMode,
+    val lastValidationsList: ArrayList<Validation>? = null,
+    val titlesList: ArrayList<Contract>,
+    val errorTitle: String? = null,
+    val errorMessage: String? = null
+)
