+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/domain/model/Location.kt
@@ -1,0 +1,7 @@
+package org.calypsonet.keyple.demo.control.domain.model
+
+data class Location(val id: Int, val name: String) {
+    override fun toString(): String {
+        return "$id - $name"
+    }
+}
